# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Specialist APP APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get a listing of the Transactions.",
                "description": "Get all Transactions",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Transaction"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Store a newly created Transaction in storage",
                "description": "Store Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Transaction that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Transaction"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/transactions/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Display the specified Transaction",
                "description": "Get Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Transaction",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Transaction"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update the specified Transaction in storage",
                "description": "Update Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Transaction",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Transaction that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Transaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Transaction"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Remove the specified Transaction from storage",
                "description": "Delete Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Transaction",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/transactions/checkin": {
            "post": {
                "tags": [
                    "Mobile-Api"
                ],
                "summary": "Store a newly created Transaction in storage",
                "description": "Store Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "plate_img",
                        "in": "formData",
                        "description": "plate photo",
                        "required": true,
                        "type": "file"
                    },
                    {
                        "name": "plate_number",
                        "in": "query",
                        "description": "plate_number",
                        "required": false,
                        "type": "string",
                        "default": "sameh"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/CheckIn"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/transactions/checkin-client/{client_id}": {
            "post": {
                "tags": [
                    "Mobile-Api"
                ],
                "summary": "Store a newly client created Transaction in storage",
                "description": "Store client Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "description": "id of clinet",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/CheckIn"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/transactions/checkout/{qr_code}": {
            "post": {
                "tags": [
                    "Mobile-Api"
                ],
                "summary": "Set checkout for Transaction",
                "description": "checkout Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "qr_code",
                        "in": "path",
                        "description": "qr code of Transaction",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/CheckOut"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/transactions/checkout-clinet/{client_id}": {
            "post": {
                "tags": [
                    "Mobile-Api"
                ],
                "summary": "Set checkout for Transaction",
                "description": "checkout Transaction",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "client_id",
                        "in": "path",
                        "description": "qr code of Transaction",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/CheckOut"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "CheckIn": {
            "required": [
                "plate_number"
            ],
            "properties": {
                "plate_number": {
                    "description": "car plate number",
                    "type": "string"
                },
                "plate_img": {
                    "description": "image of car plate",
                    "type": "string"
                },
                "mobile": {
                    "description": "driver mobile",
                    "type": "string"
                },
                "driver_name": {
                    "description": "driver name",
                    "type": "string"
                }
            }
        },
...
        "Transaction": {
            "required": [
                "plate_number",
                "plate_img"
            ],
            "properties": {
                "plate_number": {
                    "description": "plate_number",
                    "type": "string"
                },
                "plate_img": {
                    "description": "plate_img",
                    "type": "string"
                },
                "out_at": {
                    "description": "out_at",
                    "type": "string"
                },
                "qr_code": {
                    "description": "qr_code",
                    "type": "string"
                },
                "is_payed": {
                    "description": "is_payed",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "description": "Auth Bearer Token\n *      Format as 'Bearer <access_token>'",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}